Rename dicts file
